pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9.9",
    "validators>=0.28.0"
]
requires-python = ">=3.12"

[tool.poetry.dependencies]
python = "^3.12"

[project.optional-dependencies]
dev = [
    "flake8"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"